+++ labs/lab8/src/main/java/edu/mum/cs545/game/services/IPlayerService.java
@@ -3,8 +3,8 @@ package edu.mum.cs545.game.services;
 import edu.mum.cs545.game.domain.Player;
 
 public interface IPlayerService {
-    Player getById(int id);
+    Player getById(Long id);
     Iterable<Player> listAll();
     Player save(Player player);
-    void delete(int id);
+    void delete(Long id);
 }
