+++ labs/lab8/src/main/java/edu/mum/cs545/game/services/IStadiumService.java
@@ -3,9 +3,9 @@ package edu.mum.cs545.game.services;
 import edu.mum.cs545.game.domain.Stadium;
 
 public interface IStadiumService {
-    Stadium getById(int id);
+    Stadium getById(Long id);
     Iterable<Stadium> listAll();
     Stadium save(Stadium stadium);
-    void delete(int id);
+    void delete(Long id);
 }
