+++ labs/lab8/src/main/java/edu/mum/cs545/game/services/ITeamService.java
@@ -3,8 +3,8 @@ package edu.mum.cs545.game.services;
 import edu.mum.cs545.game.domain.Team;
 
 public interface ITeamService {
-    Team getById(int id);
+    Team getById(Long id);
     Iterable<Team> listAll();
     Team save(Team team);
-    void delete(int id);
+    void delete(Long id);
 }
